adapted prefetching logic to be cleaner

+++ src/Servebolt/Utils/EnvFile/Reader.php
@@ -12,6 +12,7 @@ use function Servebolt\Optimizer\Helpers\getOption;
 use function Servebolt\Optimizer\Helpers\isHostedAtServebolt;
 use function Servebolt\Optimizer\Helpers\updateOption;
 use function Servebolt\Optimizer\Helpers\isNextGen;
+use function Servebolt\Optimizer\Helpers\isTesting;
 
 /**
  * Class Reader
@@ -146,9 +147,8 @@ class Reader
             return $pathFromCache;
         }
 
-        // Check path cache matches current directory structure
-        // if it does not match, we clear path cache to force a re-check.
-        if( isset($_SERVER['DOCUMENT_ROOT']) && strpos($_SERVER['DOCUMENT_ROOT'], $pathFromCache) === false) {
+        // When not it testing mode. Automatically clear the filepath if it gets this far.
+        if( !isTesting() && isset($_SERVER['DOCUMENT_ROOT']) && strpos($_SERVER['DOCUMENT_ROOT'], $pathFromCache) === false) {
             error_log('[Servebolt Optimizer] Is automatically clearing the environment filepath to force an update.');
             deleteOption($this->optionsKey);
         }   
@@ -159,8 +159,8 @@ class Reader
             $this->setResolvedFileType($pathFromDiskLookup);
             return $pathFromDiskLookup;
         }
-
-        error_log('[Servebolt Optimizer] Could not resolve environment file on the path: ' . $pathFromDiskLookup );
+        // If we get this far, we could not resolve the environment file as its missing.
+        error_log('[Servebolt Optimizer] Could not find the environment.[ini,json] file on the path: ' . $pathFromDiskLookup );
 
         return false;
     }
